requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
pillow==10.4.0
diffusers==0.27.2
transformers==4.40.0
safetensors==0.4.2
accelerate==0.29.3
xformers==0.0.26
python-multipart==0.0.9
